linux-mainline: add patch to change kworker priority
This patch changes the priority of the kworker threads to be realtime
threads with a priority of 50. This will reduce the latency of the UART
and allows to detect the recuperation mode faster.


(cherry picked from commit 02fe875d092dc43c81a9e5d0f2340e1b9d0bcd4b)