Fill in hardware timestamp field in skb of flexcan with monotonic ktime.
For logging purposes it makes sense to have monotonic timestamps in candump.
Since Flexcan does not have hardware timestamps, the predefined ts[2] of skb
is used to hold the value of ktime_get() for both RX and TX.
This "fake" hardware timestamp can then be accessed by candump via the -H
option which will then use SO_TIMESTAMPING and obtain this value.

With having NTP in mind to realize a mobile connectivity watchdog in future
and possibly shortening the PollInterval of systemd-timesyncd, using a
monotonic timestamp for candump becomes even more attractive.