Set "spi-rx-bus-width = <0>;" for st7789v to operate it with SPI_NO_RX.
With https://github.com/torvalds/linux/commit/290cdd7959a734a0ef20ec096af7810177c4b9f8
a change was introduced to read and check the display ID and print a warning
if the check fails as a method to speed up bring-ups.
Since the panel was operated with no MISO line so far and "spi-3wire;" does not
work out of the box, it seems reasonable to disable the check by setting
"spi-rx-bus-width = <0>;".